% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/File.R
\name{dpih}
\alias{dpih}
\title{Find the best "Der Preis ist heiß" answer}
\usage{
dpih(x = "Objekt")
}
\arguments{
\item{x}{A product (character)}
}
\value{
The price of the product
}
\description{
Find the best "Der Preis ist heiß" answer
}
\examples{
dpih("How much is the fish?")
}

