% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/File.R
\name{ezod}
\alias{ezod}
\title{Find the best "1, 2 oder 3" answer}
\usage{
ezod(x = "Frage")
}
\arguments{
\item{x}{A question (character)}
}
\value{
The number of the best answer
}
\description{
Find the best "1, 2 oder 3" answer
}
\examples{
ezod("Which of the three is correct?")
}

