% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/File.R
\name{gag}
\alias{gag}
\title{Find the best "Geh aufs Ganze" answer}
\usage{
gag(x = "Tor")
}
\arguments{
\item{x}{The kind of choice (character)}
}
\value{
The definitive best choice
}
\description{
Find the best "Geh aufs Ganze" answer
}
\examples{
gag("Umschlag")
gag("Tor")
}

