% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/File.R
\name{wwm}
\alias{wwm}
\title{Find the best "Wer wird Millionaer" answer}
\usage{
wwm(x = "Frage")
}
\arguments{
\item{x}{A question (character)}
}
\value{
The letter of the best answer
}
\description{
Find the best "Wer wird Millionaer" answer
}
\examples{
wwm("Which of the four is correct?")
}

